Mieczy

+++ src/main/java/org/cmccx/utils/S3Service.java
@@ -18,7 +18,6 @@ import java.util.List;
 
 @RequiredArgsConstructor
 @Service
-
 public class S3Service {
     final Logger logger = LoggerFactory.getLogger(this.getClass());
 
@@ -82,11 +81,13 @@ public class S3Service {
     /** 파일 삭제 **/
     public void deleteImage(String imageUrl) throws BaseException {
         try {
-            // 파일명 디코딩
-            String fileName = imageUrl.substring(imageUrl.lastIndexOf('/') + 1, imageUrl.length());
-            String decodeData = URLDecoder.decode(fileName, "UTF-8");
-            // 파일 삭제
-            amazonS3.deleteObject(new DeleteObjectRequest(bucket, directory + decodeData));
+            if (imageUrl != null) {
+                // 파일명 디코딩
+                String fileName = imageUrl.substring(imageUrl.lastIndexOf('/') + 1, imageUrl.length());
+                String decodeData = URLDecoder.decode(fileName, "UTF-8");
+                // 파일 삭제
+                amazonS3.deleteObject(new DeleteObjectRequest(bucket, directory + decodeData));
+            }
         } catch (Exception e) {
             logger.error("S3 ERROR", e);
             throw new BaseException(BaseResponseStatus.S3_DELETE_ERROR);
